Cleanup and docs
* move RLDM IPLD schema in common
* clean some namings
* add docs
