+++ packages/repco-common/lib.ts
@@ -1,3 +1,4 @@
 export * as zod from './src/zod.js'
 export * from './src/log.js'
 export * from './src/util.js'
+export * as schema from './src/schema.js'
