+++ packages/repco-prisma/prisma/schema.prisma
@@ -5,7 +5,7 @@ datasource db {
 
 generator client {
   provider        = "prisma-client-js"
-  previewFeatures = ["interactiveTransactions"]
+  previewFeatures = ["deno"]
 }
 
 // generator dbml {
